chore: remove unnused stuff and move arround folders
