+++ packages/api/src/http/routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { setupHotelRoutes } from "./hotels.routes";
+import { setupHotelRoutes } from "./hotels/hotels.routes";
 
 export async function setupRoutes() {
   const router = Router();
