+++ packages/api/src/types/common.ts
@@ -1,5 +1,4 @@
 import type { NextFunction, Request, Response } from "express";
-import { z } from "zod";
 
 export interface Context {
   req: Request;
